+++ BLCMM/src/blcmm/gui/panels/EditPanel.java
@@ -117,7 +117,7 @@ public class EditPanel extends javax.swing.JPanel implements InputValidator, Can
         this.patchType.getLevels().forEach(map -> levelnames.add(map.code));
 
         initComponents();
-        textElement = new HighlightedTextArea(true, allowEdit);
+        textElement = new HighlightedTextArea(MainGUI.INSTANCE.getDMM(), true, allowEdit);
         jPanel2.setLayout(new BorderLayout());
         jPanel2.add(textElement);
         jScrollPane1.getVerticalScrollBar().setUnitIncrement(16);