+++ BLCMM/src/CHANGELOG.md
@@ -52,8 +52,16 @@ BLCMM Changelog
 - "Restore defaults" button on settings screen will only restore the settings
   shown on the active screen (this prevents clearing the recently-viewed file
   list, for instance)
-- Object autocomplete will correctly handle `:` separators in addition to
-  periods.
+- Autocomplete changes:
+  - Object autocomplete will correctly handle `:` separators in addition to
+    periods.
+  - Fixed some problems with full-object-name substitution in the main code
+    edit area.
+  - Attribute/Field autocompletes will pull from the entire pool of top-level
+    attribute names
+  - Attribute value autocompletes will pull from the entire pool of Enum
+    values
+
 - Removed hidden ability to save in legacy FilterTool mod format
 
 (changelog entries to confirm, etc -- from original 1.2.1 plans ages ago)